install script improvements

+++ install.sh
@@ -5,13 +5,13 @@ SCRIPTPATH=$(dirname "$(realpath "$0")")
 DOTFILES_BACKUP_DIR="$HOME/.dotfiles.old"
 mkdir -p "$DOTFILES_BACKUP_DIR"
 
-cp "$HOME/.bash_aliases" "$DOTFILES_BACKUP_DIR/"
-cp "$HOME/.bash_completion" "$DOTFILES_BACKUP_DIR/"
-cp "$HOME/.bash_functions" "$DOTFILES_BACKUP_DIR/"
-cp "$HOME/.bash_profile" "$DOTFILES_BACKUP_DIR/"
-cp "$HOME/.bashrc" "$DOTFILES_BACKUP_DIR/"
-cp "$HOME/.gitconfig" "$DOTFILES_BACKUP_DIR/"
-cp "$HOME/.vimrc" "$DOTFILES_BACKUP_DIR/"
+[ -f "$HOME/.bash_aliases" ] && cp "$HOME/.bash_aliases" "$DOTFILES_BACKUP_DIR/"
+[ -f "$HOME/.bash_completion" ] && cp "$HOME/.bash_completion" "$DOTFILES_BACKUP_DIR/"
+[ -f "$HOME/.bash_functions" ] && cp "$HOME/.bash_functions" "$DOTFILES_BACKUP_DIR/"
+[ -f "$HOME/.bash_profile" ] && cp "$HOME/.bash_profile" "$DOTFILES_BACKUP_DIR/"
+[ -f "$HOME/.bashrc" ] && cp "$HOME/.bashrc" "$DOTFILES_BACKUP_DIR/"
+[ -f "$HOME/.gitconfig" ] && cp "$HOME/.gitconfig" "$DOTFILES_BACKUP_DIR/"
+[ -f "$HOME/.vimrc" ] && cp "$HOME/.vimrc" "$DOTFILES_BACKUP_DIR/"
 
 ln -s -f "$SCRIPTPATH/.bash_aliases" $HOME/.bash_aliases
 ln -s -f "$SCRIPTPATH/.bash_completion" $HOME/.bash_completion
